seperate db logic to individual files
